changes to avoid message error when it does not fin both multiplicites in the excite states set

+++ manager.sh
@@ -122,20 +122,24 @@ else
         
 	for ii in 0 1 # repeating same process for multiplicity 0 and 1
 	do
-		out2_step4="trans${ii}_st.out"
-                ./step8.sh $out2_step4 $exc_range $opt_soc
-	        mv corevirtMO_matrix.out corevirtMO_matrix${ii}.out
-	        mv corevirtMO_matrix_ts_probability.out corevirtMO_matrix_ts_probability${ii}.out
-	        mv corevirtMO_matrix.csv corevirtMO_matrix${ii}.csv
-	        mv corevirtMO_matrix_ts_probability.csv corevirtMO_matrix_ts_probability${ii}.csv
+		if [[ -f "trans${ii}_st.out" ]]; then
+			# It can be the case when excited state does not have one of the
+			# multiplicities
+		        out2_step4="trans${ii}_st.out"
+                        ./step8.sh $out2_step4 $exc_range $opt_soc
+	                mv corevirtMO_matrix.out corevirtMO_matrix${ii}.out
+	                mv corevirtMO_matrix_ts_probability.out corevirtMO_matrix_ts_probability${ii}.out
+	                mv corevirtMO_matrix.csv corevirtMO_matrix${ii}.csv
+	                mv corevirtMO_matrix_ts_probability.csv corevirtMO_matrix_ts_probability${ii}.csv
 	
-	        # Additional step to differentiate which 9th step to perform
-                if (( $spectra==0 )); then
-                	# Updated step9 version which only works with corrected spectra version
-                        ./step9_soc.sh $out2_step4 $out_file $opt_soc
-                fi
-                mv corevirt_fosc_corr_matrix.csv corevirt_fosc_corr_matrix${ii}.csv
-                mv corevirt_foscw_corr_matrix.csv corevirt_foscw_corr_matrix${ii}.csv
+	                # Additional step to differentiate which 9th step to perform
+                        if (( $spectra==0 )); then
+                        	# Updated step9 version which only works with corrected spectra version
+                                ./step9_soc.sh $out2_step4 $out_file $opt_soc
+                        fi
+                        mv corevirt_fosc_corr_matrix.csv corevirt_fosc_corr_matrix${ii}.csv
+                        mv corevirt_foscw_corr_matrix.csv corevirt_foscw_corr_matrix${ii}.csv
+		fi
 	done
 	#mv corevirtMO_matrix.out corevirtMO_matrix0.out
 	#mv corevirtMO_matrix_ts_probability.out corevirtMO_matrix_ts_probability0.out
@@ -165,21 +169,21 @@ cp ${out_file}_out/resB_MOcore.csv pop_matrices/${folder2}/resB_MOvirt_${pop_nam
 
 if (( $opt_soc!=1 )); then
 
-	cp ${out_file}_out/corevirtMO_matrix.csv pop_matrices/${folder2}/corevirtMO_matrix_${pop_name}_${sufff}.csv
-        cp ${out_file}_out/corevirtMO_matrix_ts_probability.csv pop_matrices/${folder2}/corevirtMO_matrix_tspb_${pop_name}_${sufff}.csv
-	cp ${out_file}_out/corevirt_fosc_e_matrix.csv pop_matrices/${folder2}/corevirt_fosc_${pop_name}_${sufff}.csv
-	cp ${out_file}_out/corevirt_fosc_we_matrix.csv pop_matrices/${folder2}/corevirt_foscw_${pop_name}_${sufff}.csv
+	cp ${out_file}_out/corevirtMO_matrix.csv pop_matrices/${folder2}/corevirtMO_matrix_${pop_name}_${sufff}.csv 2>/dev/null
+        cp ${out_file}_out/corevirtMO_matrix_ts_probability.csv pop_matrices/${folder2}/corevirtMO_matrix_tspb_${pop_name}_${sufff}.csv 2>/dev/null
+	cp ${out_file}_out/corevirt_fosc_e_matrix.csv pop_matrices/${folder2}/corevirt_fosc_${pop_name}_${sufff}.csv 2>/dev/null
+	cp ${out_file}_out/corevirt_fosc_we_matrix.csv pop_matrices/${folder2}/corevirt_foscw_${pop_name}_${sufff}.csv 2>/dev/null
 
 else
 	
 	for ii in 0 1
 	do
-		cp ${out_file}_out/corevirtMO_matrix${ii}.csv pop_matrices/${folder2}/corevirtMO_matrix${ii}_${pop_name}_${sufff}.csv
-                cp ${out_file}_out/corevirtMO_matrix_ts_probability${ii}.csv pop_matrices/${folder2}/corevirtMO_matrix${ii}_tspb_${pop_name}_${sufff}.csv
+		cp ${out_file}_out/corevirtMO_matrix${ii}.csv pop_matrices/${folder2}/corevirtMO_matrix${ii}_${pop_name}_${sufff}.csv 2>/dev/null
+                cp ${out_file}_out/corevirtMO_matrix_ts_probability${ii}.csv pop_matrices/${folder2}/corevirtMO_matrix${ii}_tspb_${pop_name}_${sufff}.csv 2>/dev/null
 
 		# Additional step only for the SOC evaluation
-		cp ${out_file}_out/corevirt_fosc_corr_matrix${ii}.csv pop_matrices/${folder2}/corevirt_fosc${ii}_corr_${pop_name}_${sufff}.csv
-	        cp ${out_file}_out/corevirt_foscw_corr_matrix${ii}.csv pop_matrices/${folder2}/corevirt_foscw${ii}_corr_${pop_name}_${sufff}.csv
+		cp ${out_file}_out/corevirt_fosc_corr_matrix${ii}.csv pop_matrices/${folder2}/corevirt_fosc${ii}_corr_${pop_name}_${sufff}.csv 2>/dev/null
+	        cp ${out_file}_out/corevirt_foscw_corr_matrix${ii}.csv pop_matrices/${folder2}/corevirt_foscw${ii}_corr_${pop_name}_${sufff}.csv 2>/dev/null
 	done
 
 fi
